+++ e2e/auth.e2e.ts
@@ -1,22 +1,24 @@
 import { providers, Wallet } from 'ethers';
 import nock from 'nock';
 import {
+  AuthService,
   AuthTokens,
   CacheClient,
   ProviderType,
   setCacheConfig,
   SignerService,
-  TEST_LOGIN_ENDPOINT,
+  DEFAULT_AUTH_STATUS_PATH,
 } from '../src';
 import { rpcUrl } from './utils/setup-contracts';
 
 describe('Authentication tests', () => {
-  const SSI_HUB_URL = 'http://localhost:8080';
+  const SSI_HUB_URL = 'http://identitycache.energyweb.org/v1';
   const provider = new providers.JsonRpcProvider({ url: rpcUrl });
-
-  const loginPathBody = { identityToken: /(^[\w-]*\.[\w-]*\.[\w-]*$)/ };
+  const domain = 'http://localhost:8080';
+  const RETRY_COUNT = 3; // number of retried requests
 
   let signerService: SignerService;
+  let authService: AuthService;
   let cacheClient: CacheClient;
 
   const getNockScope = (): nock.Scope => nock(SSI_HUB_URL);
@@ -25,7 +27,10 @@ describe('Authentication tests', () => {
     const network = await provider.getNetwork();
     setCacheConfig(network.chainId, {
       url: SSI_HUB_URL,
-      cacheServerSupportsAuth: true,
+      auth: {
+        domain,
+        baseUrl: SSI_HUB_URL,
+      },
     });
   });
 
@@ -37,6 +42,8 @@ describe('Authentication tests', () => {
     );
     cacheClient = new CacheClient(signerService);
     await signerService.init();
+
+    authService = cacheClient['authService'];
   });
 
   afterEach(() => {
@@ -46,16 +53,12 @@ describe('Authentication tests', () => {
   describe('isAuthenticated()', () => {
     let nockScope: nock.Scope;
 
-    beforeEach(() => {
-      nockScope = getNockScope();
-    });
-
     afterEach(() => {
       expect(nockScope.isDone()).toBe(true);
     });
 
     it('should return true if the user is authenticated', async () => {
-      nockScope.get(TEST_LOGIN_ENDPOINT).reply(200, {
+      nockScope = getNockScope().get(DEFAULT_AUTH_STATUS_PATH).reply(200, {
         user: signerService.did,
       });
 
@@ -64,7 +67,7 @@ describe('Authentication tests', () => {
     });
 
     it('should return false if other user is authenticated', async () => {
-      nockScope.get(TEST_LOGIN_ENDPOINT).reply(200, {
+      nockScope = getNockScope().get(DEFAULT_AUTH_STATUS_PATH).reply(200, {
         user: 'did:ethr:0x0000000000000000000000000000000000000000',
       });
 
@@ -73,7 +76,7 @@ describe('Authentication tests', () => {
     });
 
     it('should return false if empty user in response', async () => {
-      nockScope.get(TEST_LOGIN_ENDPOINT).reply(200, {
+      nockScope = getNockScope().get(DEFAULT_AUTH_STATUS_PATH).reply(200, {
         user: null,
       });
 
@@ -82,7 +85,7 @@ describe('Authentication tests', () => {
     });
 
     it('should return false if server error occurred', async () => {
-      nockScope.get(TEST_LOGIN_ENDPOINT).reply(500);
+      nockScope = getNockScope().get(DEFAULT_AUTH_STATUS_PATH).reply(500);
 
       const isAuthenticated = await cacheClient.isAuthenticated();
       expect(isAuthenticated).toBe(false);
@@ -92,10 +95,6 @@ describe('Authentication tests', () => {
   describe('refreshToken()', () => {
     let nockScope: nock.Scope;
 
-    beforeEach(() => {
-      nockScope = getNockScope();
-    });
-
     it('should obtain new tokens from auth server', async () => {
       const newTokens = {
         token: 'new-token',
@@ -103,15 +102,19 @@ describe('Authentication tests', () => {
       };
       const oldRefreshToken = 'old-token';
 
-      nockScope
+      nockScope = getNockScope()
         .get(`/refresh_token?refresh_token=${oldRefreshToken}`)
         .reply(200, newTokens);
 
-      cacheClient['refresh_token'] = oldRefreshToken;
+      authService['refresh_token'] = oldRefreshToken;
 
-      await expect(cacheClient['refreshToken']()).resolves.toStrictEqual(
-        newTokens
+      await authService['refreshToken']();
+      expect(authService['refresh_token']).toStrictEqual(
+        newTokens.refreshToken
       );
+      expect(
+        cacheClient['_httpClient'].defaults.headers.common.Authorization
+      ).toBe(`Bearer ${newTokens.token}`);
       expect(nockScope.isDone()).toBe(true);
     });
 
@@ -126,46 +129,56 @@ describe('Authentication tests', () => {
         .get(`/refresh_token?refresh_token=${oldRefreshToken}`)
         .reply(200, newTokens);
 
-      await expect(cacheClient['refreshToken']()).resolves.toBeUndefined();
+      await expect(authService['refreshToken']()).resolves.toBeUndefined();
       expect(nockScope.isDone()).toBe(false);
     });
 
     it('should throw axios error if occurred', async () => {
       const oldRefreshToken = 'old-token';
 
-      nockScope
+      nockScope = getNockScope()
         .get(`/refresh_token?refresh_token=${oldRefreshToken}`)
+        .times(RETRY_COUNT + 1)
         .reply(500);
 
-      cacheClient['refresh_token'] = oldRefreshToken;
+      authService['refresh_token'] = oldRefreshToken;
 
-      await expect(cacheClient['refreshToken']()).rejects.toBeDefined();
+      await expect(authService['refreshToken']()).rejects.toBeDefined();
       expect(nockScope.isDone()).toBe(true);
     });
   });
 
   describe('login()', () => {
-    it('should not authenticated when client is already authenticated', async () => {
-      cacheClient['isAuthenticated'] = jest.fn().mockResolvedValueOnce(true);
-      cacheClient['authenticate'] = jest.fn();
+    it('should not authenticate when client is already authenticated', async () => {
+      authService['isAuthenticated'] = jest.fn().mockResolvedValueOnce(true);
+      authService['authenticate'] = jest.fn();
       await cacheClient.login();
-      expect(cacheClient['authenticate']).not.toHaveBeenCalled();
+      expect(authService['authenticate']).not.toHaveBeenCalled();
     });
 
     it('should authenticate when client is not authenticated yet', async () => {
-      cacheClient['isAuthenticated'] = jest.fn().mockResolvedValueOnce(false);
-      cacheClient['authenticate'] = jest.fn();
+      authService['isAuthenticated'] = jest.fn().mockResolvedValueOnce(false);
+      authService['authenticate'] = jest.fn();
       await cacheClient.login();
-      expect(cacheClient['authenticate']).toHaveBeenCalledTimes(1);
+      expect(authService['authenticate']).toHaveBeenCalledTimes(1);
     });
   });
 
   describe('authenticate()', () => {
+    let authInitScope: nock.Scope;
+    const nonce = 'i3cNDECwQD7JvpWW5';
+
+    beforeEach(() => {
+      authInitScope = getNockScope()
+        .post('/login/siwe/initiate')
+        .reply(201, { nonce });
+    });
+
     const checkTokens = (tokens: AuthTokens) => {
       expect(
         cacheClient['_httpClient'].defaults.headers.common.Authorization
       ).toBe(`Bearer ${tokens.token}`);
-      expect(cacheClient['refresh_token']).toBe(tokens.refreshToken);
+      expect(authService['refresh_token']).toBe(tokens.refreshToken);
     };
 
     it('should refresh tokens', async () => {
@@ -175,24 +188,27 @@ describe('Authentication tests', () => {
       };
       const oldRefreshToken = 'old-token';
 
-      cacheClient['refresh_token'] = oldRefreshToken;
+      authService['refresh_token'] = oldRefreshToken;
 
       const refreshScope = getNockScope()
         .get(`/refresh_token?refresh_token=${oldRefreshToken}`)
         .reply(200, newTokens);
 
       const loginScope = getNockScope()
-        .post(`/login`, loginPathBody)
+        .post(`/login/siwe/verify`)
         .reply(200, newTokens);
 
-      const statusScope = getNockScope().get(TEST_LOGIN_ENDPOINT).reply(200, {
-        user: signerService.did,
-      });
+      const statusScope = getNockScope()
+        .get(DEFAULT_AUTH_STATUS_PATH)
+        .reply(200, {
+          user: signerService.did,
+        });
 
       await cacheClient.authenticate();
 
       checkTokens(newTokens);
 
+      expect(authInitScope.isDone()).toBe(false);
       expect(refreshScope.isDone()).toBe(true);
       expect(loginScope.isDone()).toBe(false);
       expect(statusScope.isDone()).toBe(true);
@@ -205,22 +221,31 @@ describe('Authentication tests', () => {
       };
 
       const refreshScope = getNockScope()
-        .get(/^\/refresh_token/)
+        .get('refresh_token')
         .reply(200, newTokens);
 
       const loginScope = getNockScope()
-        .post(`/login`, loginPathBody)
+        .post(`/login/siwe/verify`)
         .reply(200, newTokens);
 
-      const statusScope = getNockScope().get(TEST_LOGIN_ENDPOINT).reply(200, {
-        user: signerService.did,
-      });
+      const statusScope = getNockScope()
+        .get(DEFAULT_AUTH_STATUS_PATH)
+        .once()
+        .reply(200, {
+          user: null,
+        })
+        .get(DEFAULT_AUTH_STATUS_PATH)
+        .once()
+        .reply(200, {
+          user: signerService.did,
+        });
 
       await cacheClient.authenticate();
 
       checkTokens(newTokens);
 
       expect(refreshScope.isDone()).toBe(false);
+      expect(authInitScope.isDone()).toBe(true);
       expect(loginScope.isDone()).toBe(true);
       expect(statusScope.isDone()).toBe(false);
     });
@@ -233,25 +258,34 @@ describe('Authentication tests', () => {
 
       const oldRefreshToken = 'old-token';
 
-      cacheClient['refresh_token'] = oldRefreshToken;
+      authService['refresh_token'] = oldRefreshToken;
 
       const refreshScope = getNockScope()
-        .get(/^\/refresh_token/)
-        .reply(500, newTokens);
+        .get(/refresh_token/)
+        .reply(500);
 
       const loginScope = getNockScope()
-        .post(`/login`, loginPathBody)
+        .post('/login/siwe/verify')
         .reply(200, newTokens);
 
-      const statusScope = getNockScope().get(TEST_LOGIN_ENDPOINT).reply(200, {
-        user: signerService.did,
-      });
+      const statusScope = getNockScope()
+        .get(DEFAULT_AUTH_STATUS_PATH)
+        .once()
+        .reply(200, {
+          user: null,
+        })
+        .get(DEFAULT_AUTH_STATUS_PATH)
+        .once()
+        .reply(200, {
+          user: signerService.did,
+        });
 
       await cacheClient.authenticate();
 
       checkTokens(newTokens);
 
       expect(refreshScope.isDone()).toBe(true);
+      expect(authInitScope.isDone()).toBe(true);
       expect(loginScope.isDone()).toBe(true);
       expect(statusScope.isDone()).toBe(false);
     });
@@ -264,120 +298,96 @@ describe('Authentication tests', () => {
 
       const oldRefreshToken = 'old-token';
 
-      cacheClient['refresh_token'] = oldRefreshToken;
+      authService['refresh_token'] = oldRefreshToken;
 
       const refreshScope = getNockScope()
-        .get(/^\/refresh_token/)
+        .get(/refresh_token/)
         .reply(200, newTokens);
 
       const loginScope = getNockScope()
-        .post(`/login`, loginPathBody)
+        .post('/login/siwe/verify')
         .reply(200, newTokens);
 
-      const statusScope = getNockScope().get(TEST_LOGIN_ENDPOINT).reply(200, {
-        user: 'did:ethr:volta:0x0000000000000000000000000000000000000000',
-      });
+      const statusScope = getNockScope()
+        .get(DEFAULT_AUTH_STATUS_PATH)
+        .reply(200, {
+          user: 'did:ethr:volta:0x0000000000000000000000000000000000000000',
+        });
 
       await cacheClient.authenticate();
 
       checkTokens(newTokens);
 
       expect(refreshScope.isDone()).toBe(true);
+      expect(authInitScope.isDone()).toBe(true);
       expect(loginScope.isDone()).toBe(true);
       expect(statusScope.isDone()).toBe(true);
     });
 
     it('should throw an error when login fails', async () => {
-      const newTokens = {
-        token: 'new-token',
-        refreshToken: 'new-refresh-token',
-      };
-
-      const loginScope = getNockScope()
-        .post(`/login`, loginPathBody)
-        .reply(500, newTokens);
+      const loginScope = getNockScope().post(`/login/siwe/verify`).reply(500);
 
       await expect(cacheClient.authenticate()).rejects.toBeDefined();
 
+      expect(authInitScope.isDone()).toBe(true);
       expect(loginScope.isDone()).toBe(true);
     });
   });
 
   describe('makeRetryRequest()', () => {
-    const MOCK_REQUEST_PATH = '/api/v1/test';
-    const mockRequest = jest.fn().mockImplementation(async () => {
-      const { data } = await cacheClient['_httpClient'].get<{
-        success: boolean;
-      }>(MOCK_REQUEST_PATH);
-
-      if (!data.success) {
-        throw new Error('Request failed');
-      }
-
-      return data;
-    });
-
-    afterEach(() => {
-      mockRequest.mockClear();
-    });
+    const roleName = 'test';
+    const ROLE_DEF_PATH = `/role/${roleName}`;
+    const roleData = { definition: { roleName: 'test' } };
 
     it('should result with data after successful request', async () => {
-      const nockScope = getNockScope().get(MOCK_REQUEST_PATH).reply(200, {
-        success: true,
-      });
+      const nockScope = getNockScope().get(ROLE_DEF_PATH).reply(200, roleData);
 
-      const data = await cacheClient['makeRetryRequest'](mockRequest);
+      const def = await cacheClient.getRoleDefinition(roleName);
 
-      expect(data).toEqual({ success: true });
-      expect(mockRequest).toHaveBeenCalledTimes(1);
+      expect(def).toEqual(roleData.definition);
       expect(nockScope.isDone()).toBe(true);
     });
 
-    it('should not retry not axios error', async () => {
-      const nockScope = getNockScope().get(MOCK_REQUEST_PATH).reply(200, {
-        success: false,
-      });
-
-      await expect(
-        cacheClient['makeRetryRequest'](mockRequest)
-      ).rejects.toThrow('Request failed');
-
-      expect(mockRequest).toHaveBeenCalledTimes(1);
-      expect(nockScope.isDone()).toBe(true);
-    });
-
-    it('should retry 5xx errors', async () => {
+    // unclear how to mock non axios error
+    it.skip('should not retry not axios error', async () => {
+      const errorMessage = 'Request failed';
       const nockScope = getNockScope()
-        .get(MOCK_REQUEST_PATH)
+        .get(ROLE_DEF_PATH)
+        .replyWithError(errorMessage);
+
+      await expect(cacheClient.getRoleDefinition(roleName)).rejects.toThrow(
+        errorMessage
+      );
+
+      expect(nockScope.isDone()).toBe(true);
+    });
+
+    it('should retry 500 errors', async () => {
+      const nockScope = getNockScope()
+        .get(ROLE_DEF_PATH)
         .reply(500)
-        .get(MOCK_REQUEST_PATH)
-        .reply(200, {
-          success: true,
-        });
+        .get(ROLE_DEF_PATH)
+        .reply(200, roleData);
 
-      const data = await cacheClient['makeRetryRequest'](mockRequest);
+      const data = await cacheClient.getRoleDefinition(roleName);
 
-      expect(data).toEqual({ success: true });
-      expect(mockRequest).toHaveBeenCalledTimes(2);
+      expect(data).toEqual(roleData.definition);
       expect(nockScope.isDone()).toBe(true);
     });
 
     it('should retry ECONNREFUSED error', async () => {
       const nockScope = getNockScope()
-        .get(MOCK_REQUEST_PATH)
+        .get(ROLE_DEF_PATH)
         .replyWithError({
           code: 'ECONNREFUSED',
           errno: 'ECONNREFUSED',
         })
-        .get(MOCK_REQUEST_PATH)
-        .reply(200, {
-          success: true,
-        });
+        .get(ROLE_DEF_PATH)
+        .reply(200, roleData);
 
-      const data = await cacheClient['makeRetryRequest'](mockRequest);
+      const def = await cacheClient.getRoleDefinition(roleName);
 
-      expect(data).toEqual({ success: true });
-      expect(mockRequest).toHaveBeenCalledTimes(2);
+      expect(def).toEqual(roleData.definition);
       expect(nockScope.isDone()).toBe(true);
     });
 
@@ -385,17 +395,14 @@ describe('Authentication tests', () => {
       408, 411, 412, 425, 426, 500, 501, 502, 503, 504, 505, 506, 510, 511,
     ])('should retry %i error', async (statusCode) => {
       const nockScope = getNockScope()
-        .get(MOCK_REQUEST_PATH)
+        .get(ROLE_DEF_PATH)
         .reply(statusCode)
-        .get(MOCK_REQUEST_PATH)
-        .reply(200, {
-          success: true,
-        });
+        .get(ROLE_DEF_PATH)
+        .reply(200, roleData);
 
-      const data = await cacheClient['makeRetryRequest'](mockRequest);
+      const def = await cacheClient.getRoleDefinition(roleName);
 
-      expect(data).toEqual({ success: true });
-      expect(mockRequest).toHaveBeenCalledTimes(2);
+      expect(def).toEqual(roleData.definition);
       expect(nockScope.isDone()).toBe(true);
     });
 
@@ -403,149 +410,141 @@ describe('Authentication tests', () => {
       400, 402, 404, 405, 406, 409, 410, 413, 414, 415, 416, 417, 422, 428, 429,
       431, 451,
     ])('should not retry %i error', async (statusCode) => {
-      const nockScope = getNockScope().get(MOCK_REQUEST_PATH).reply(statusCode);
+      const nockScope = getNockScope().get(ROLE_DEF_PATH).reply(statusCode);
 
-      await expect(
-        cacheClient['makeRetryRequest'](mockRequest)
-      ).rejects.toThrow();
+      await expect(cacheClient.getRoleDefinition(roleName)).rejects.toThrow();
 
-      expect(mockRequest).toHaveBeenCalledTimes(1);
       expect(nockScope.isDone()).toBe(true);
     });
 
     it('should not retry auth endpoints', async () => {
-      const nockScope = getNockScope().post('/login').reply(401);
+      const nockScope = getNockScope().post('/login/siwe/initiate').reply(401);
 
-      const authMockRequest = jest.fn().mockImplementation(async () => {
-        return await cacheClient['_httpClient'].post<{
-          success: boolean;
-        }>('/login');
-      });
+      await expect(cacheClient.authenticate()).rejects.toThrow();
 
-      await expect(
-        cacheClient['makeRetryRequest'](authMockRequest)
-      ).rejects.toThrow();
-
-      expect(authMockRequest).toHaveBeenCalledTimes(1);
       expect(nockScope.isDone()).toBe(true);
     });
 
-    it('should authenticate when 401 error occurred', async () => {
-      const nockScope = getNockScope()
-        .get(MOCK_REQUEST_PATH)
+    it('should authenticate when responds with 401 status code', async () => {
+      const roleDefScope = getNockScope()
+        .get(ROLE_DEF_PATH)
         .reply(401)
-        .get(MOCK_REQUEST_PATH)
-        .reply(200, {
-          success: true,
-        });
+        .get(ROLE_DEF_PATH)
+        .reply(200, roleData);
 
-      cacheClient['authenticate'] = jest.fn().mockImplementation(async () => {
-        return await new Promise((resolve) => {
-          setTimeout(() => {
-            resolve(0);
-          }, 100);
-        });
-      });
+      const authStatusScope = getNockScope()
+        .get('/auth/status')
+        .reply(200, { user: null });
+      const nonce = 'i3cNDECwQD7JvpWW5';
+      const authInitScope = getNockScope()
+        .post('/login/siwe/initiate')
+        .reply(200, { nonce });
+      const tokens: AuthTokens = {
+        token: 'access token',
+        refreshToken: 'refresh token',
+      };
+      const authVefifyScope = getNockScope()
+        .post('/login/siwe/verify')
+        .reply(200, tokens);
 
-      const data = await cacheClient['makeRetryRequest'](mockRequest);
+      const def = await cacheClient.getRoleDefinition(roleName);
 
-      expect(data).toEqual({ success: true });
-      expect(mockRequest).toHaveBeenCalledTimes(2);
-      expect(nockScope.isDone()).toBe(true);
-      expect(cacheClient['authenticate']).toHaveBeenCalledTimes(1);
+      expect(def).toEqual(roleData.definition);
+      expect(roleDefScope.isDone()).toBe(true);
+      expect(authInitScope.isDone()).toBe(true);
+      expect(authStatusScope.isDone()).toBe(true);
+      expect(authVefifyScope.isDone()).toBe(true);
     });
 
-    it('should authenticate once when 401 error occurred in other request during ongoing authentication process', async () => {
-      const nockScope = getNockScope()
-        .get(MOCK_REQUEST_PATH)
+    it('should authenticate once on two unauthenticated requests', async () => {
+      const roleDefScope = getNockScope()
+        .get(ROLE_DEF_PATH)
         .reply(401)
-        .get(MOCK_REQUEST_PATH)
-        .reply(200, {
-          success: true,
-        });
+        .get(ROLE_DEF_PATH)
+        .reply(200, roleData);
 
-      const mockOtherRequestPath = '/api/v1/test2';
-      const nockScope2 = getNockScope()
-        .get(mockOtherRequestPath)
+      const orgName = 'test';
+      const ORG_DEF_PATH = `/org/${orgName}`;
+      const orgData = { definition: { orgName: 'Org name' } };
+      const orgDefScope = getNockScope()
+        .get(ORG_DEF_PATH)
         .reply(401)
-        .get(mockOtherRequestPath)
-        .reply(200, {
-          success: true,
-        });
+        .get(ORG_DEF_PATH)
+        .reply(200, orgData);
 
-      const mockOtherRequest = jest.fn().mockImplementation(async () => {
-        const { data } = await cacheClient['_httpClient'].get<{
-          success: boolean;
-        }>(mockOtherRequestPath);
-        return data;
-      });
-
-      cacheClient['authenticate'] = jest.fn().mockImplementation(async () => {
-        return await new Promise((resolve) => {
-          setTimeout(() => {
-            resolve(0);
-          }, 1000);
-        });
-      });
+      const nonce = 'i3cNDECwQD7JvpWW5';
+      const authInitScope = getNockScope()
+        .post('/login/siwe/initiate')
+        .reply(200, { nonce });
+      const tokens: AuthTokens = {
+        token: 'access token',
+        refreshToken: 'refresh token',
+      };
+      const authVefifyScope = getNockScope()
+        .post('/login/siwe/verify')
+        .reply(200, tokens);
 
       const data = await Promise.all([
-        cacheClient['makeRetryRequest'](mockRequest),
-        cacheClient['makeRetryRequest'](mockOtherRequest),
+        cacheClient.getRoleDefinition(roleName),
+        cacheClient.getOrgDefinition(orgName),
       ]);
 
       expect(data).toStrictEqual(
-        expect.arrayContaining([{ success: true }, { success: true }])
+        expect.arrayContaining([roleData.definition, orgData.definition])
       );
-      expect(mockRequest).toHaveBeenCalledTimes(2);
-      expect(mockOtherRequest).toHaveBeenCalledTimes(2);
-      expect(nockScope.isDone()).toBe(true);
-      expect(nockScope2.isDone()).toBe(true);
-      expect(cacheClient['authenticate']).toHaveBeenCalledTimes(1);
+      expect(roleDefScope.isDone()).toBe(true);
+      expect(orgDefScope.isDone()).toBe(true);
+      expect(authInitScope.isDone()).toBe(true);
+      expect(authVefifyScope.isDone()).toBe(true);
     });
 
-    it('should retry when handler request threw error', async () => {
-      const nockScope = getNockScope()
-        .get(MOCK_REQUEST_PATH)
+    it('should retry when request handler throw error', async () => {
+      const roleDefScope = getNockScope()
+        .get(ROLE_DEF_PATH)
         .reply(500)
-        .get(MOCK_REQUEST_PATH)
-        .reply(200, {
-          success: true,
-        });
+        .get(ROLE_DEF_PATH)
+        .reply(200, roleData);
 
-      cacheClient['handleRequestError'] = jest.fn().mockRejectedValue('error');
+      jest
+        .spyOn(authService as any, 'handleRequestError')
+        .mockRejectedValueOnce(new Error());
 
-      const data = await cacheClient['makeRetryRequest'](mockRequest);
+      const def = await cacheClient.getRoleDefinition(roleName);
 
-      expect(data).toEqual({ success: true });
-      expect(mockRequest).toHaveBeenCalledTimes(2);
-      expect(nockScope.isDone()).toBe(true);
+      expect(def).toEqual(roleData.definition);
+      expect(roleDefScope.isDone()).toBe(true);
     });
 
-    it('should retry when auth endpoint threw 5xx error', async () => {
-      const requestScope = getNockScope()
-        .get(MOCK_REQUEST_PATH)
+    it('should retry when login verify endpoint responds with 500 status code', async () => {
+      const getRoleScope = getNockScope()
+        .get(ROLE_DEF_PATH)
         .reply(401)
-        .get(MOCK_REQUEST_PATH)
-        .reply(401)
-        .get(MOCK_REQUEST_PATH)
-        .reply(200, {
-          success: true,
-        });
+        .get(ROLE_DEF_PATH)
+        .reply(200, roleData);
 
-      const loginScope = getNockScope()
-        .post('/login', loginPathBody)
+      const authStatusScope = getNockScope()
+        .get('/auth/status')
+        .reply(200, { user: null });
+      const authInitScope = getNockScope()
+        .post('/login/siwe/initiate')
+        .reply(201, { nonce: 'i3cNDECwQD7JvpWW5' });
+      const tokens: AuthTokens = {
+        token: 'access token',
+        refreshToken: 'refresh token',
+      };
+      const authVerifyScope = getNockScope()
+        .post('/login/siwe/verify')
         .reply(500)
-        .post('/login', loginPathBody)
-        .reply(201, {
-          success: true,
-        });
+        .post('/login/siwe/verify')
+        .reply(201, tokens);
 
-      const data = await cacheClient['makeRetryRequest'](mockRequest);
+      const def = await cacheClient.getRoleDefinition(roleName);
 
-      expect(data).toEqual({ success: true });
-      expect(mockRequest).toHaveBeenCalledTimes(3);
-      expect(requestScope.isDone()).toBe(true);
-      expect(loginScope.isDone()).toBe(true);
+      expect(def).toEqual(roleData.definition);
+      expect(getRoleScope.isDone()).toBe(true);
+      expect(authInitScope.isDone()).toBe(true);
+      expect(authVerifyScope.isDone()).toBe(true);
+      expect(authStatusScope.isDone()).toBe(true);
     });
   });
 });