+++ src/modules/cache-client/cache-client.types.ts
@@ -1,4 +1,5 @@
 import { AssetHistoryEventType } from '../assets';
+import { SiweOptions } from '../auth';
 
 export enum Order {
   'ASC' = 'ASC',
@@ -8,6 +9,7 @@ export enum Order {
 export interface CacheServerClientOptions {
   url: string;
   cacheServerSupportsAuth?: boolean;
+  auth: SiweOptions;
 }
 
 export type ClaimsFilter = {
@@ -26,11 +28,4 @@ export enum SearchType {
   App = 'App',
   Org = 'Org',
   Role = 'Role',
-}
-
-export const TEST_LOGIN_ENDPOINT = '/auth/status';
-
-export interface AuthTokens {
-  token: string;
-  refreshToken: string;
 }